feat: RESTful API interfacing Private Blockchain
This commit includes all changes required to complete the project.

1) The app.js file

This file defines the class for the REST API.

2) The BlockController.js file

This file defines the controller to encapsulate routes to work with blocks. It features a GET Endpoint to retrieve a block by index, url: "/block/:index". In addition, it also includes a POST Endpoint to add a new Block, url: "/block". Furthermore, it contains a helper method to inizialize a mock dataset, adding 10 test blocks to the blockchain, on top of the genesis block already created.

3) The levelSandbox.js file

This file contains the Data Access Layer for the application. The private blockchain to be created uses LevelDB for persistance. Both project dependencies and storage location are indicated in this file.

4) The Block.js file

This file contains the Block class with a constructor for block. All the attributes that make up a block are included.

5) The BlockChain.js file

This file contains the Blockchain class with a constructor for blockchain.

6) The README.md file

This file documents the project.